Track Finding Cleanup, main branch (2025.01.07.) (#808)
* Common sense cleanup in the device track finding.

Introduced the globalIndex_t typedef for the unsigned int
type. Switched all track finding functions to using it
instead of std::size_t.

Removed all unnecessary templating from functions that
were never updated after the track finding configuration
became a simple standalone struct.

Removed some unnecessary const declarations from the
payload structs, and added some extra const declarations
in some of the function bodies.

* CUDA updates for the common track finding changes.

* Reduce the number of parallel tests in GitLab CI.

Running too many CUDA tests in parallel requires a
*lot* of memory. This will hopefully make the tests
a bit more reliable.

* globalIndex_t -> global_index_t

+++ device/cuda/src/finding/finding_algorithm.cu
@@ -1,6 +1,6 @@
 /** TRACCC library, part of the ACTS project (R&D line)
  *
- * (c) 2023-2024 CERN for the benefit of the ACTS project
+ * (c) 2023-2025 CERN for the benefit of the ACTS project
  *
  * Mozilla Public License Version 2.0
  */
@@ -189,8 +189,8 @@ finding_algorithm<stepper_t, navigator_t>::operator()(
 
             kernels::apply_interaction<std::decay_t<detector_type>>
                 <<<nBlocks, nThreads, 0, stream>>>(
-                    m_cfg, {det_view, static_cast<int>(n_in_params),
-                            in_params_buffer, param_liveness_buffer});
+                    m_cfg, {det_view, n_in_params, in_params_buffer,
+                            param_liveness_buffer});
             TRACCC_CUDA_ERROR_CHECK(cudaGetLastError());
         }
